+++ xchange-core/src/main/java/org/knowm/xchange/service/account/params/AccountLeverageParams.java
@@ -1,0 +1,5 @@
+package org.knowm.xchange.service.account.params;
+
+public interface AccountLeverageParams {
+    int getLeverage();
+}
