+++ xchange-ftx/src/main/java/org/knowm/xchange/ftx/service/FtxAccountServiceRaw.java
@@ -1,10 +1,8 @@
 package org.knowm.xchange.ftx.service;
 
 import java.io.IOException;
-import java.net.URLEncoder;
 import java.util.List;
 import org.knowm.xchange.Exchange;
-import org.knowm.xchange.dto.account.AccountLeverageSetting;
 import org.knowm.xchange.ftx.FtxException;
 import org.knowm.xchange.ftx.dto.FtxResponse;
 import org.knowm.xchange.ftx.dto.account.*;
@@ -122,19 +120,15 @@ public class FtxAccountServiceRaw extends FtxBaseService {
     }
   }
 
-  /**
-   * @deprecated use {@link FtxAccountService#setInitialLeverage(AccountLeverageSetting)}
-   */
-  @Deprecated
-  public FtxResponse<FtxLeverageDto> changeLeverage(String subaccount, int leverage)
-      throws FtxException, IOException {
+  public FtxResponse<FtxLeverageDto> setLeverage(String subaccount, int leverage)
+          throws FtxException, IOException {
     try {
       return ftx.changeLeverage(
-          exchange.getExchangeSpecification().getApiKey(),
-          exchange.getNonceFactory().createValue(),
-          signatureCreator,
-          null,
-          new FtxLeverageDto(leverage));
+              exchange.getExchangeSpecification().getApiKey(),
+              exchange.getNonceFactory().createValue(),
+              signatureCreator,
+              subaccount,
+              new FtxLeverageDto(leverage));
     } catch (FtxException e) {
       throw new FtxException(e.getMessage());
     }
